web/requirements.txt

# RFID-I2C Web Sunucusu Gereksinimleri

# Web Framework
Flask==2.3.3
Flask-SocketIO==5.3.6

# MQTT İstemci
paho-mqtt==1.6.1

# WebSocket Desteği
python-socketio==5.9.0
python-engineio==4.7.1

# JSON İşleme (Python ile birlikte gelir ama belirtmek için)
# json - Built-in

# Tarih/Saat İşlemleri (Python ile birlikte gelir)
# datetime - Built-in

# Threading Desteği (Python ile birlikte gelir)  
# threading - Built-in

# Ek Geliştirme Araçları (isteğe bağlı)
# python-dotenv==1.0.0  # Çevre değişkenleri için
# gunicorn==21.2.0      # Production sunucu için